chore: Fix flakey test (#202)
