+++ tests/e2e/pipeline.go
@@ -1,0 +1,148 @@
+package e2e
+
+import (
+	"fmt"
+
+	. "github.com/onsi/gomega"
+
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+	"k8s.io/client-go/util/retry"
+
+	numaflowv1 "github.com/numaproj/numaflow/pkg/apis/numaflow/v1alpha1"
+
+	"github.com/numaproj/numaplane/internal/util"
+	"github.com/numaproj/numaplane/internal/util/kubernetes"
+	apiv1 "github.com/numaproj/numaplane/pkg/apis/numaplane/v1alpha1"
+)
+
+func verifyPipelineSpec(namespace string, pipelineName string, f func(numaflowv1.PipelineSpec) bool) {
+
+	document("verifying Pipeline Spec")
+	var retrievedPipelineSpec numaflowv1.PipelineSpec
+	Eventually(func() bool {
+		unstruct, err := dynamicClient.Resource(getGVRForPipeline()).Namespace(namespace).Get(ctx, pipelineName, metav1.GetOptions{})
+		if err != nil {
+			return false
+		}
+		if retrievedPipelineSpec, err = getPipelineSpec(unstruct); err != nil {
+			return false
+		}
+
+		return f(retrievedPipelineSpec)
+	}).WithTimeout(testTimeout).Should(BeTrue())
+}
+
+func verifyPipelineStatus(namespace string, pipelineName string, f func(numaflowv1.PipelineSpec, kubernetes.GenericStatus) bool) {
+
+	document("verifying PipelineStatus")
+	var retrievedPipelineSpec numaflowv1.PipelineSpec
+	var retrievedPipelineStatus kubernetes.GenericStatus
+	Eventually(func() bool {
+		unstruct, err := dynamicClient.Resource(getGVRForPipeline()).Namespace(namespace).Get(ctx, pipelineName, metav1.GetOptions{})
+		if err != nil {
+			return false
+		}
+		if retrievedPipelineSpec, err = getPipelineSpec(unstruct); err != nil {
+			return false
+		}
+		if retrievedPipelineStatus, err = getNumaflowResourceStatus(unstruct); err != nil {
+			return false
+		}
+
+		return f(retrievedPipelineSpec, retrievedPipelineStatus)
+	}).WithTimeout(testTimeout).Should(BeTrue())
+}
+
+func verifyPipelineReady(namespace string, pipelineName string, numVertices int) {
+	document("Verifying that the Pipeline is running")
+	verifyPipelineStatus(namespace, pipelineName,
+		func(retrievedPipelineSpec numaflowv1.PipelineSpec, retrievedPipelineStatus kubernetes.GenericStatus) bool {
+			return retrievedPipelineStatus.Phase == string(numaflowv1.PipelinePhaseRunning)
+		})
+
+	vertexLabelSelector := fmt.Sprintf("%s=%s,%s=%s", numaflowv1.KeyPipelineName, pipelineName, numaflowv1.KeyComponent, "vertex")
+	daemonLabelSelector := fmt.Sprintf("%s=%s,%s=%s", numaflowv1.KeyPipelineName, pipelineName, numaflowv1.KeyComponent, "daemon")
+
+	// Get Pipeline Pods to verify they're all up
+	document("Verifying that the Pipeline is ready")
+	// check "vertex" Pods
+	verifyPodsRunning(namespace, 2, vertexLabelSelector)
+	verifyPodsRunning(namespace, 1, daemonLabelSelector)
+
+}
+
+// Get PipelineSpec from Unstructured type
+func getPipelineSpec(u *unstructured.Unstructured) (numaflowv1.PipelineSpec, error) {
+	specMap := u.Object["spec"]
+	var pipelineSpec numaflowv1.PipelineSpec
+	err := util.StructToStruct(&specMap, &pipelineSpec)
+	return pipelineSpec, err
+}
+
+func getGVRForPipeline() schema.GroupVersionResource {
+	return schema.GroupVersionResource{
+		Group:    "numaflow.numaproj.io",
+		Version:  "v1alpha1",
+		Resource: "pipelines",
+	}
+}
+
+func updatePipelineRolloutInK8S(namespace string, name string, f func(apiv1.PipelineRollout) (apiv1.PipelineRollout, error)) {
+	document("updating PipelineRollout")
+	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
+		rollout, err := pipelineRolloutClient.Get(ctx, name, metav1.GetOptions{})
+		if err != nil {
+			return err
+		}
+
+		*rollout, err = f(*rollout)
+		if err != nil {
+			return err
+		}
+		_, err = pipelineRolloutClient.Update(ctx, rollout, metav1.UpdateOptions{})
+		return err
+	})
+	Expect(err).ShouldNot(HaveOccurred())
+}
+
+func updatePipelineSpecInK8S(namespace string, pipelineName string, f func(numaflowv1.PipelineSpec) (numaflowv1.PipelineSpec, error)) {
+	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
+
+		unstruct, err := dynamicClient.Resource(getGVRForPipeline()).Namespace(namespace).Get(ctx, pipelineName, metav1.GetOptions{})
+		Expect(err).ShouldNot(HaveOccurred())
+		retrievedPipeline := unstruct
+
+		// modify Pipeline Spec to verify it gets auto-healed
+		err = updatePipelineSpec(retrievedPipeline, f)
+		Expect(err).ShouldNot(HaveOccurred())
+		_, err = dynamicClient.Resource(getGVRForPipeline()).Namespace(namespace).Update(ctx, retrievedPipeline, metav1.UpdateOptions{})
+		return err
+	})
+	Expect(err).ShouldNot(HaveOccurred())
+}
+
+// Take a Pipeline Unstructured type and update the PipelineSpec in some way
+func updatePipelineSpec(u *unstructured.Unstructured, f func(numaflowv1.PipelineSpec) (numaflowv1.PipelineSpec, error)) error {
+
+	// get PipelineSpec from unstructured object
+	specMap := u.Object["spec"]
+	var pipelineSpec numaflowv1.PipelineSpec
+	err := util.StructToStruct(&specMap, &pipelineSpec)
+	if err != nil {
+		return err
+	}
+	// update PipelineSpec
+	pipelineSpec, err = f(pipelineSpec)
+	if err != nil {
+		return err
+	}
+	var newMap map[string]interface{}
+	err = util.StructToStruct(&pipelineSpec, &newMap)
+	if err != nil {
+		return err
+	}
+	u.Object["spec"] = newMap
+	return nil
+}
